export CollapsibleAssetCard and AssetCard as CommonAssetCard and CommonCard
